Rewrite the epsilon workchain and fix example
